Avoid executing commands twice on Windows
fmt.Scanf separates input by space or newline. Since Windows command
prompt input ends in "\n\r", fmt.Scanf parses the actual input, then the
for-loop runs again and it sees the "\r" as another input. This triggers
an error "unexpected newline", which is ignored, and this.execute() gets
called again with no change to this.instruction. fmt.Scanln doesn't have
this problem.